pyproject.toml
[project]
name = "openteddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.4.12",
    "httpx>=0.28.1",
    "lancedb>=0.30.1",
    "langchain-openai>=1.1.12",
    "openai>=2.30.0",
    "opensandbox-code-interpreter>=0.1.2",
    "opensandbox-server>=0.1.9",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
]
